Add: HelpDialogの入力を変更できるように
