+++ src/components/UpdateInfo.vue
@@ -6,9 +6,7 @@
     <div class="q-pa-md">
       <template v-if="isUpdateAvailable">
         <h3>最新バージョン {{ latestVersion }} が見つかりました</h3>
-        <a href="https://voicevox.hiroshiba.jp/" target="_blank"
-          >ダウンロードページ</a
-        >
+        <a href="{{ downloadLink }}" target="_blank">ダウンロードページ</a>
         <hr />
       </template>
       <h3>アップデート履歴</h3>
@@ -35,69 +33,28 @@
 
 <script lang="ts">
 import { useStore } from "@/store";
-import { computed, defineComponent, ref } from "@vue/runtime-core";
+import { computed, defineComponent, PropType, ref } from "@vue/runtime-core";
 import { UpdateInfo } from "../type/preload";
 import semver from "semver";
 
 export default defineComponent({
-  setup() {
-    const store = useStore();
-
-    const updateInfos = ref<UpdateInfo[]>();
-    store.dispatch("GET_UPDATE_INFOS").then((obj) => (updateInfos.value = obj));
-
-    let isCheckingFinished = ref<boolean>(false);
-
-    // 最新版があるか調べる
-    const currentVersion = ref("");
-    const latestVersion = ref("");
-    window.electron
-      .getAppInfos()
-      .then((obj) => {
-        currentVersion.value = obj.version;
-      })
-      .then(() => {
-        fetch("https://api.github.com/repos/VOICEVOX/voicevox/releases", {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-            Accept: "application/json",
-          },
-        })
-          .then((response) => {
-            if (!response.ok) throw new Error("Network response was not ok.");
-            return response.json();
-          })
-          .then((json) => {
-            const newerVersion = json.find(
-              (item: { prerelease: boolean; tag_name: string }) => {
-                return (
-                  !item.prerelease &&
-                  semver.valid(currentVersion.value) &&
-                  semver.valid(item.tag_name) &&
-                  semver.lt(currentVersion.value, item.tag_name)
-                );
-              }
-            );
-            if (newerVersion) {
-              latestVersion.value = newerVersion.tag_name;
-            }
-            isCheckingFinished.value = true;
-          })
-          .catch((err) => {
-            throw new Error(err);
-          });
-      });
-
-    const isUpdateAvailable = computed(() => {
-      return isCheckingFinished.value && latestVersion.value !== "";
-    });
-
-    return {
-      isUpdateAvailable,
-      latestVersion,
-      updateInfos,
-    };
+  props: {
+    latestVersion: {
+      type: String,
+      required: false,
+    },
+    downloadLink: {
+      type: String,
+      required: false,
+    },
+    updateInfos: {
+      type: Array as PropType<UpdateInfo[]>,
+      required: false,
+    },
+    isUpdateAvailable: {
+      type: Boolean,
+      required: false,
+    },
   },
 });
 </script>
